refactor: Move the utils in respective directories for sorting and preferences
