+++ src/components/TopologyPopover/topologySort.tsx
@@ -1,7 +1,98 @@
 import clsx from "clsx";
 import { BsSortDown, BsSortUp } from "react-icons/bs";
 
-import { defaultSortLabels } from "./utils";
+import { isDate } from "../../utils/date";
+
+import type { Topology, ValueType } from "../../context/TopologyPageContext";
+
+const STATUS = {
+  info: 0,
+  healthy: 1,
+  warning: 2,
+  unhealthy: 3
+};
+
+export const defaultSortLabels = [
+  { id: 1, value: "status", label: "Health" },
+  { id: 2, value: "name", label: "Name" },
+  { id: 3, value: "type", label: "Type" },
+  { id: 4, value: "updated_at", label: "Last Updated" }
+];
+
+export function getSortLabels(topology: Topology[]) {
+  const currentSortLabels: typeof defaultSortLabels = [];
+
+  topology?.forEach((t) => {
+    t?.properties?.forEach((h, index) => {
+      if (h.headline && !currentSortLabels.find((t) => t.value === h.name)) {
+        currentSortLabels.push({
+          id: defaultSortLabels.length + index,
+          value: (h.name ?? '').toLowerCase(),
+          label: h.name ?? ''
+        });
+      }
+    });
+  });
+
+  return [...defaultSortLabels, ...currentSortLabels];
+}
+
+function getTopologyValue(t: Topology, sortBy: string) {
+  if (Boolean(t[sortBy])) {
+    return t[sortBy] as ValueType;
+  }
+
+  const property = t?.properties?.find((p) => p.name === sortBy);
+  if (property) {
+    return property.value as ValueType;
+  }
+
+  return undefined;
+}
+
+export function getSortedTopology(
+  topology: Topology[] = [],
+  sortBy: string,
+  sortByType: string
+) {
+  const topologyMap = new Map(topology.map((p) => [p.id, p]));
+
+  let updatedTopology = [...topologyMap.values()].sort((t1, t2) => {
+    let t1Value = getTopologyValue(t1, sortBy);
+    let t2Value = getTopologyValue(t2, sortBy);
+
+    if (t1Value && (!t2Value || t2Value === null)) {
+      return 1;
+    }
+    if (t2Value && (!t1Value || t1Value === null)) {
+      return -1;
+    }
+
+    if (isDate(t1Value) && isDate(t2Value)) {
+      return (
+        new Date(t1Value as string).getDate() -
+        new Date(t2Value as string).getDate()
+      );
+    }
+
+    if (sortBy === "status") {
+      t1Value = STATUS[t1Value];
+      t2Value = STATUS[t2Value];
+    }
+
+    if (t1Value && t2Value) {
+      return +(t1Value > t2Value) || -(t1Value < t2Value);
+    }
+
+    return 0;
+  });
+
+  if (sortByType === "desc") {
+    return updatedTopology.reverse();
+  }
+
+  return updatedTopology;
+}
 
 export const TopologySort = ({
   title = "Sort By",
