+++ src/utils/date.ts
@@ -1,0 +1,13 @@
+import { ValueType } from "../context/TopologyPageContext";
+
+export function isDate(dateString?: ValueType) {
+  if (
+    !dateString ||
+    dateString === null ||
+    !["string", "object"].includes(typeof dateString)
+  ) {
+    return false;
+  }
+
+  return !isNaN(new Date(dateString).getDate());
+}
